backend/requirements.txt
fastapi[all]
langchain 
sqlalchemy 
hypothesis 
pypdf2 
pymupdf
openpyxl 
google-genai 
groq
pgvector
supabase
streamlit 
requests 
reportlab 
supabase